requirements.txt
# ==== Core Web Framework ====
Flask==3.0.3
Werkzeug==3.1.3
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.2.1
blinker==1.9.0
colorama==0.4.6
tqdm==4.66.5
requests==2.31.0
typing_extensions==4.13.2
filelock==3.18.0
packaging==25.0
regex==2024.11.6
urllib3==2.4.0
idna==3.10
certifi==2025.4.26
numpy==1.26.3
scipy==1.15.3
sympy==1.14.0
mpmath==1.3.0
networkx==3.4.2
sentence-transformers==3.3.1
transformers==4.46.3
huggingface-hub==0.26.5
safetensors==0.5.3
nltk==3.9.1
sentencepiece==0.2.0
faiss-cpu==1.8.0
pillow==11.2.1
PyYAML==6.0.2
fsspec==2025.5.1
joblib==1.5.1
scikit-learn==1.6.1
threadpoolctl==3.6.0
MarkupSafe==3.0.2

pdfplumber 
pillow 
pytesseract
